# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: University Placement Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /postings:
    get:
      operationId: listPostings
      tags: [postings]
      summary: List postings (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, closed]
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of postings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Posting"
    post:
      operationId: createPosting
      tags: [postings]
      summary: Create a new posting (company only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posting"
  /postings/{id}:
    get:
      operationId: getPosting
      tags: [postings]
      summary: Get a single posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posting details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostingDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePosting
      tags: [postings]
      summary: Update a posting (company only, pending state)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostingUpdate"
      responses:
        "200":
          description: Updated posting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posting"
  /postings/{id}/approve:
    patch:
      operationId: approvePosting
      tags: [postings]
      summary: Approve a posting (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posting approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posting"
  /postings/{id}/reject:
    patch:
      operationId: rejectPosting
      tags: [postings]
      summary: Reject a posting (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionInput"
      responses:
        "200":
          description: Posting rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posting"
  /postings/{id}/close:
    patch:
      operationId: closePosting
      tags: [postings]
      summary: Close a posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posting closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posting"
...
components:
  schemas:
...
    Posting:
      type: object
      required: [id, companyId, companyName, title, description, status, deadline, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        title:
          type: string
        description:
          type: string
        eligibility:
          type: ["string", "null"]
        minCgpa:
          type: ["number", "null"]
        branches:
          type: ["string", "null"]
        ctc:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        slots:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, closed]
        rejectionReason:
          type: ["string", "null"]
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        applicationCount:
          type: ["integer", "null"]
        hasApplied:
          type: ["boolean", "null"]
    PostingDetail:
      type: object
      required: [id, companyId, companyName, title, description, status, deadline, createdAt, applications]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        title:
          type: string
        description:
          type: string
        eligibility:
          type: ["string", "null"]
        minCgpa:
          type: ["number", "null"]
        branches:
          type: ["string", "null"]
        ctc:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        slots:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, closed]
        rejectionReason:
          type: ["string", "null"]
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
    PostingInput:
      type: object
      required: [title, description, deadline]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        eligibility:
          type: string
        minCgpa:
          type: number
        branches:
          type: string
        ctc:
          type: string
        location:
          type: string
        slots:
          type: integer
        deadline:
          type: string
          format: date-time
    PostingUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        eligibility:
          type: string
        minCgpa:
          type: number
        branches:
          type: string
        ctc:
          type: string
        location:
          type: string
        slots:
          type: integer
        deadline:
          type: string
          format: date-time
    RejectionInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
    Application:
      type: object
      required: [id, postingId, studentId, status, appliedAt]
      properties:
        id:
          type: integer
        postingId:
          type: integer
        postingTitle:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        studentEmail:
          type: ["string", "null"]
        studentDepartment:
          type: ["string", "null"]
        studentCgpa:
          type: ["number", "null"]
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected]
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...