# lib/api-zod/src/generated/types/postingDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * University Placement Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';
import type { PostingDetailStatus } from './postingDetailStatus';

export interface PostingDetail {
  id: number;
  companyId: number;
  companyName: string;
  title: string;
  description: string;
  /** @nullable */
  eligibility?: string | null;
  /** @nullable */
  minCgpa?: number | null;
  /** @nullable */
  branches?: string | null;
  /** @nullable */
  ctc?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  slots?: number | null;
  status: PostingDetailStatus;
  /** @nullable */
  rejectionReason?: string | null;
  deadline: Date;
  createdAt: Date;
  updatedAt?: Date;
  applications: Application[];
}