# lib/api-zod/src/generated/types/postingUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * University Placement Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface PostingUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  eligibility?: string;
  minCgpa?: number;
  branches?: string;
  ctc?: string;
  location?: string;
  slots?: number;
  deadline?: Date;
}