# lib/api-zod/src/generated/types/rejectionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * University Placement Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface RejectionInput {
  /** @minLength 1 */
  reason: string;
}